requirements.txt
supabase
psycopg2-binary
python-dotenv
fastapi
uvicorn
pydantic
httpx
gunicorn
